requirements.txt
filelock==3.15.4
fsspec==2024.6.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
pillow==10.4.0
PyYAML==6.0.2
setuptools==72.2.0
sympy==1.13.2
torch==2.4.0
torchsummary==1.5.1
torchvision==0.19.0
tqdm==4.66.5
typing_extensions==4.12.2
